action.download moduleを作成
'_download_thread.py'の行数･class数が増加したため,
DownloadThread関連のclassをaction.download moduleへ移動した.
