+++ slackbot/action/__init__.py
@@ -2,10 +2,4 @@
 
 from ._api_logger import APILogger
 from ._download import Download
-from ._download_thread import (
-    DownloadException,
-    DownloadReport,
-    DownloadReportType,
-    DownloadThread,
-    DownloadThreadOption)
 from ._response import Response
